+++ tests/api/unit_tests/test_unit_image_export_options.py
@@ -62,6 +62,7 @@ class TestUnitImageExportOptions:
         + [("ShowHistogram", "show_histogram", x) for x in VALID_BOOL]
         + [("ShowMinMax", "show_min_max", x) for x in VALID_BOOL]
         + [("FitToScreen", "fit_to_screen", x) for x in VALID_BOOL]
+        + [("ShowTransparentBg", "transparent_background", x) for x in VALID_BOOL]
         + [("CaptureMode", "capture_mode", x.value) for x in CaptureModes],
     )
     # pylint: disable-next=R0913, R0917
@@ -106,6 +107,7 @@ class TestUnitImageExportOptions:
         + [("ShowHistogram", "show_histogram", x, x) for x in VALID_BOOL]
         + [("ShowMinMax", "show_min_max", x, x) for x in VALID_BOOL]
         + [("FitToScreen", "fit_to_screen", x, x) for x in VALID_BOOL]
+        + [("ShowTransparentBg", "transparent_background", x, x) for x in VALID_BOOL]
         + [("CaptureMode", "capture_mode", x, x.value) for x in CaptureModes],
     )
     # pylint: disable-next=R0913, R0917
@@ -148,6 +150,7 @@ class TestUnitImageExportOptions:
         + [("ShowHistogram", "show_histogram", x) for x in INVALID_BOOL]
         + [("ShowMinMax", "show_min_max", x) for x in INVALID_BOOL]
         + [("FitToScreen", "fit_to_screen", x) for x in INVALID_BOOL]
+        + [("ShowTransparentBg", "transparent_background", x) for x in INVALID_BOOL]
         + [("CaptureMode", "capture_mode", x) for x in INVALID_INT],
     )
     # pylint: disable-next=R0913, R0917